--- src/app/(provider)/(root)/js-test/page.tsx
@@ -1,11 +1,0 @@
-import Card from "@/components/Card";
-
-function TestPage() {
-  return (
-    <div className="max-w-[375px]">
-      <Card />
-    </div>
-  );
-}
-
-export default TestPage;
