+++ src/app/(provider)/(root)/recommend/page.tsx
@@ -1,0 +1,11 @@
+import IntroCard from "./_components/IntroCard";
+
+function RecommendIntroPage() {
+  return (
+    <div className="w-screen h-screen max-h-[812px] flex justify-center">
+      <IntroCard />
+    </div>
+  );
+}
+
+export default RecommendIntroPage;
